service/requirements.txt
fastapi==0.68.1
torch==1.9.0
torchvision==0.10.0
Pillow==8.3.2
python-multipart==0.0.5
aiofiles==0.7.0